py/src/requirements.txt
beautifulsoup4==4.9.3
chardet==4.0.0
html5lib==1.1
imap-tools==0.45.0
imbox==0.9.8
lxml==4.6.3
numpy==1.21.1
pandas==1.3.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
webencodings==0.5.1
